Checkin of some handy utilities used on linux.
mount.c opens the specified serial port and flushes the buffers. It then passes the file descriptors to mount.

versionsim.c opens the serial port and sends the bytes to simulate the Tversion message.
This is handy because it magically resets something inside v9fs that returns control back to the
process that issued the 9p message upon some erronious transaction.