Created 'pos' directory to test read operations on motor encoders.
Created 'motor' directory to test write operations on the motor pwm control.

Created p9_noread and p9_nowrite default functions that do nothing.

Created a simple C program to test the number of reads per second that can be
achieved at a given baud rate. This repeated read was the only operation happening
on the 9p link. A reading of 140 reads/second was achieved at 57600 baud.
This should roughly double at 115200 baud.

The EMC use case for this application is 1 read and 1 write per cycle so the above
is not realistic representation of real world use. Further testing shall ensue.